+++ src/dungeon/level_utils.rs
@@ -16,15 +16,15 @@ pub fn add_xy_to_idx(width: i32, x: i32, y: i32, idx: i32) -> i32 {
     idx + x + width * y
 }
 
-pub fn idxs_are_adjacent(width: u8, idx1: usize, idx2: usize) -> bool {
+pub fn idxs_are_adjacent(width: usize, idx1: usize, idx2: usize) -> bool {
     idx1 == idx2 + 1
         || idx1 == idx2 - 1
-        || idx1 == idx2 + width as usize
-        || idx1 == idx2 - width as usize
-        || idx1 == idx2 + 1 + width as usize
-        || idx1 == idx2 - 1 + width as usize
-        || idx1 == idx2 + 1 - width as usize
-        || idx1 == idx2 - 1 - width as usize
+        || idx1 == idx2 + width
+        || idx1 == idx2 - width
+        || idx1 == idx2 + 1 + width
+        || idx1 == idx2 - 1 + width
+        || idx1 == idx2 + 1 - width
+        || idx1 == idx2 - 1 - width
 }
 
 pub fn idx_point(width: u32, idx: usize) -> Point {
