+++ src/services/blood_spawner.rs
@@ -5,7 +5,7 @@ pub struct BloodSpawnerRequest {
     pub fg: RGB,
     pub bg: RGB,
     pub glyph: u16,
-    pub level: u8,
+    pub level: usize,
 }
 
 pub struct BloodSpawner {
@@ -19,7 +19,7 @@ impl BloodSpawner {
         }
     }
 
-    pub fn request(&mut self, idx: usize, fg: RGB, bg: RGB, glyph: u16, level: u8) {
+    pub fn request(&mut self, idx: usize, fg: RGB, bg: RGB, glyph: u16, level: usize) {
         self.requests.push(BloodSpawnerRequest {
             idx,
             fg,
