+++ src/utils.rs
@@ -24,7 +24,7 @@ pub fn get_random_element<'a, T>(rng: &mut RandomNumberGenerator, elements: &'a 
     elements.get(idx as usize).unwrap()
 }
 
-pub fn get_current_level_from_world(world: &World) -> u8 {
+pub fn get_current_level_from_world(world: &World) -> usize {
     let player_ent = world.fetch::<Entity>();
     let dungeon_level = world.read_storage::<Position>();
     dungeon_level.get(*player_ent).unwrap().level
